+++ src/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(classes)
 add_subdirectory(data)
 add_subdirectory(expression)
 add_subdirectory(genericitems)
+add_subdirectory(kernels)
 add_subdirectory(keywords)
 add_subdirectory(io)
 add_subdirectory(main)
