+++ src/kernels/energy.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2023 Team Dissolve and contributors
 
-#include "classes/energykernel.h"
+#include "kernels/energy.h"
 #include "classes/box.h"
 #include "classes/cell.h"
 #include "classes/configuration.h"
