+++ src/kernels/energy.h
