+++ src/modules/energy/process.cpp
@@ -4,8 +4,8 @@
 #include "base/lineparser.h"
 #include "base/sysfunc.h"
 #include "classes/box.h"
-#include "classes/energykernel.h"
 #include "classes/species.h"
+#include "kernels/energy.h"
 #include "main/dissolve.h"
 #include "math/regression.h"
 #include "modules/energy/energy.h"
