+++ src/modules/forces/process.cpp
@@ -4,8 +4,8 @@
 #include "base/lineparser.h"
 #include "base/sysfunc.h"
 #include "classes/box.h"
-#include "classes/forcekernel.h"
 #include "classes/species.h"
+#include "kernels/force.h"
 #include "main/dissolve.h"
 #include "modules/forces/forces.h"
 #include "modules/import_trajectory/importtraj.h"
