+++ src/modules/intrashake/process.cpp
@@ -5,12 +5,12 @@
 #include "base/sysfunc.h"
 #include "classes/box.h"
 #include "classes/changestore.h"
-#include "classes/energykernel.h"
 #include "classes/regionaldistributor.h"
 #include "classes/species.h"
 #include "classes/speciesangle.h"
 #include "classes/speciesbond.h"
 #include "classes/speciestorsion.h"
+#include "kernels/energy.h"
 #include "main/dissolve.h"
 #include "modules/intrashake/intrashake.h"
