+++ src/modules/molshake/process.cpp
@@ -5,9 +5,9 @@
 #include "base/timer.h"
 #include "classes/box.h"
 #include "classes/changestore.h"
-#include "classes/energykernel.h"
 #include "classes/regionaldistributor.h"
 #include "classes/species.h"
+#include "kernels/energy.h"
 #include "main/dissolve.h"
 #include "modules/molshake/molshake.h"
